requirements.txt
mysql_connector_repackaged==0.3.1
Pillow==10.0.0
Pillow==10.3.0
python-dotenv==1.0.1
ttkthemes==3.2.2
